post updated

+++ BACKEND/server/db/migrations/202401101134-balance_humanity_blog_post.js
@@ -8,7 +8,7 @@ module.exports = {
                 allowNull: false,
                 autoIncrement: true,
                 primaryKey: true,
-                type: DataTypes.INTEGER(100),
+                type: DataTypes.INTEGER(255),
             },
             uuid: {
                 type: DataTypes.STRING(36),
